Move the StatisticsCollector into separate package
