+++ src/main/java/sorald/FileUtils.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.json.JSONObject;
-import sorald.event.StatisticsCollector;
+import sorald.stats.StatisticsCollector;
 
 public class FileUtils {
